% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/regshape.R
\name{getDNAShapeParamValues}
\alias{getDNAShapeParamValues}
\title{get DNA shape parameter values for a provided DNA sequence string
get DNA parameters: MGW, Prot, Roll, and Helt, the length of sequence should >=6}
\usage{
getDNAShapeParamValues(seq.string)
}
\arguments{
\item{seq.string:}{A DNA sequence, or a set of DNA sequences}
}
\value{
the parameter values for MGW, Prot, Roll, and Helt, for each hexamer subsequence
of seq.string, all concatenated together as one vector (MGW parameter values first, then
Propeller twist parameter values, then roll values, and then helix twist parameter values)
}
\description{
get DNA shape parameter values for a provided DNA sequence string
get DNA parameters: MGW, Prot, Roll, and Helt, the length of sequence should >=6
}
\examples{
getDNAShapeParamValues("ACGTAGATAG")
}

