% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/regshape.R
\name{getShapeScores}
\alias{getShapeScores}
\title{get DNA shape score for a set of sequences}
\usage{
getShapeScores(x)
}
\arguments{
\item{x}{A vector of character strings, each of length L characters, composed of sequences of
"A", "C", "G", or "T" characters}
}
\value{
a vector of voting fraction scores in the range [0-1], for the DNA shape-based regulatory element classifier
}
\description{
get DNA shape score for a set of sequences
}
\examples{
getShapeScores(c("ACGTACGT","ACGTACTG","ACTTACGT","GCGTGACT"))
}

