% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/regshape.R
\name{getShapeScoresSlidingWindow}
\alias{getShapeScoresSlidingWindow}
\title{get DNA shape score for a set of sequences}
\usage{
getShapeScoresSlidingWindow(seq.chars, wlen)
}
\arguments{
\item{seq.chars:}{A character string composed of "A", "C", "G", or "T" characters}

\item{wlen:}{A positive integer indicating the window size (in bp) for computing the DNA shape
based voting fraction score for the regulatory element classifier}
}
\value{
a vector of voting fraction scores in the range [0-1], for the DNA shape-based regulatory element classifier,
at each position of the sliding window (length of the vector is nchar(seq.chars) - wlen + 1).
}
\description{
get DNA shape score for a set of sequences
}
\examples{
getShapeScoresSlidingWindow("ACGATATGAACTAGACTAGTAGAGTAGAGC", 8)
}

